+++ publish-service/src/main/java/com/ericsson/eiffel/remrem/publish/controller/RemremHomeController.java
@@ -12,24 +12,22 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 */
-package com.ericsson.eiffel.remrem.publish.cli;
+package com.ericsson.eiffel.remrem.publish.controller;
 
-import org.apache.commons.cli.MissingOptionException;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 
-public class CLIExitCodes {
-    public static int CLI_EXCEPTION=1;
-    public static int CLI_MISSING_OPTION_EXCEPTION=2;
-    public static int HANDLE_CONTENT_FAILED=3;
-    public static int HANDLE_CONTENT_FILE_NOT_FOUND_FAILED=4;
-    public static int HANDLE_CONTENT_FILE_COULD_NOT_READ_FAILED=5;
-    public static int READ_JSON_FROM_CONSOLE_FAILED=6;
+/**
+ *
+ * This class redirects the REMREM URL to swagger landing page
+ *
+ */
+@Controller
+public class RemremHomeController {
 
-
-
-    public static int getExceptionCode(Exception e) {
-        if (e instanceof MissingOptionException) {
-            return CLI_MISSING_OPTION_EXCEPTION;
+        @RequestMapping(value = "/", method = RequestMethod.GET)
+        public String home() {
+                return "redirect:/swagger-ui.html";
         }
-        return CLI_EXCEPTION;
-    }
 }